10.8

+++ Heimdal/admin/add.c
@@ -104,7 +104,7 @@ kt_add(struct add_options *opt, int argc, char **argv)
 	if (opt->hex_flag) {
 	    size_t len;
 	    void *data;
-	
+
 	    len = (strlen(opt->password_string) + 1) / 2;
 
 	    data = malloc(len);
@@ -113,7 +113,7 @@ kt_add(struct add_options *opt, int argc, char **argv)
 		goto out;
 	    }
 
-	    if (hex_decode(opt->password_string, data, len) != len) {
+	    if ((size_t)hex_decode(opt->password_string, data, len) != len) {
 		free(data);
 		krb5_warn(context, ENOMEM, "hex decode failed");
 		goto out;
